✨ feat(file-manager): add support for project-specific file uploads by introducing FileUploadOptions interface and modifying uploadFile method to accommodate projectId, enhancing flexibility in file management
